requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.26.0
requests>=2.28.0
psutil>=5.9.0
ddgs>=9.0.0
anyio>=4.0.0
